refactor!: remove dead and dangerous public API before 2.0
Deletes exported API that is unused, misleading, or dangerous, and moves
backend-only parsing helpers behind internal/. After 2.0 these could only
be removed in a 3.0 under Go's compatibility promise.

Removed from constants:
- type Network
- MAINNET
- TESTNET
- PREVIEW
- PREPROD
  The enum numbered mainnet 0, the inverse of Cardano network IDs
  (mainnet = 1, testnets = 0) that every backend constructor takes, so
  uint8(constants.MAINNET) selected a testnet. Zero references existed in
  the module. Callers pass the network ID directly.
- BlockfrostBaseUrlTestnet
  The legacy Cardano testnet was decommissioned in 2023. Use
  BlockfrostBaseUrlPreview or BlockfrostBaseUrlPreprod.

Removed from backend:
- type AddressAmount
  Accidental export with zero references; blockfrost has its own
  unexported bfAddressAmount.

Moved from backend to internal/backendutil (no longer public API):
- BoundedInt
- BoundedIntFromUint64
- ParseAssetUnit
- ParseRedeemerTag
- ParseFraction
- ParseRational
- ScriptRefFromBytes
  These normalize provider API responses and were only called by the
  in-tree backend packages. Their tests move with them.

Kept exported in backend, deliberately:
- ComputeMaxTxFee: implements the ChainContext.MaxTxFee contract, so any
  out-of-tree backend needs it.
- ValidateAdditionalUtxo: also called from apollo.go.

Behavior change:
- Apollo.GetBurns now returns the absolute quantities of assets being
  burned instead of the net mint value. It was byte-identical to
  GetMints, so callers asking for burns were handed mints. Correct burn
  semantics already existed unexported as burnRequirementValue.
